wip: cargo build successful
But doesn't `cargo run` fails on the assert's because we're comparing an
index against an fn address.

Also, `cargo test` won't compile because I'm current;u hard coding the
SM::state_fns Vector. I could fix this in the tests but the tests won't
pass because the asserts will fail for the same reason `cargo run` fails.