wip: Add lazy_static and visit
Possible two choices for handling transition_to. The problem is that
the transition_to is "parsed" after fsm1 is parsed so I can't generate
the appropriate code. The thought would be to use lazy_static to make
some global data fsm1! and transition_to! could share. That feels wrong
especially if the compiler is multi-threaded.

I looked at typestate-rs and another way is to have state functions return
the transition_to function. Then we shouldn't need a transition_to macro
and eliminate the need to "share state.

The other option I thought of was to use a visitor pattern, so added
Visit. It "works" in that I see fn's and Macros and you can share state
between macros as you're implementing the Visit Trait on a struct:

 struct Visitor {
   // some state to share
 }

 impl<'ast> Visit<'ast> for Visitor {
   // Override desired functions such as:
   fn visit_item_fn(&mut self, node &'ast ItemFn) {
      // do something
   }
 }