feat: Change how transitions are done
Previously there was a fn transition_to and a proc_macro transition_to,
but this required that there be a mechanism to that the proc_macro
transition_to would have access to the state functions. Initially I was
thinking of using a global lazy_static to allow the information to be
accessed. But that seemed ugly as hell. I then looked at using the visitor
pattern available in syn, the visitor pattern will is a reasonable
approach, but then I discovered a better solution.

I looked at typestate-rs and saw that they were returning the next state
omstead of using some type of function or proc_macro. Thinking about it
this eliminated problem of having to have an explicit "transition_to"
as it was just the return value of the state function. This simplifies the
fsm1 macro and the visitor pattern would not be necessary, although I
still might use it in the future.

Further more, and possibly even more importantly, it eliminates an
ambiguity when the "transition_to" is executed multiple times in a state
function. As a return value it guarantees there is only one "transition"
per state.