feat: Adding iniitalizing SM::state_fns field
This is not working if I use Self::#process_fn_name but
Self::initial does work. It appears the problem is that
in Self::#process_fn_name the #process_fn_name is interpreted as
a "Literal" by the quote! macro but its a it's Self::initial then
initial is an "Indent" and it compiles and runs!