feat: Use VisitMut in implementing transion_to
Since the transition_to proc_macro, transition_to!, takes a function name
as the parameter. But the transition_to function in Fsm1 takes an index
into SM::state_fns. We have to mutate transition_to proc_macro changing the
name to an index. This is accomplished using the VisitMut trait after
parsing the fsm1 content.